+++ modules/tags.js
@@ -5,8 +5,7 @@ const Augur = require("augurbot-ts"),
   fs = require('fs'),
   axios = require('axios'),
   config = require('../config/config.json'),
-  u = require("../utils/utils"),
-  tu = require("../utils/tagUtils");
+  u = require("../utils/utils");
 
 /**
  * @param {Discord.Attachment} attachment
@@ -22,19 +21,65 @@ async function saveAttachment(attachment, cmd) {
   response.data.pipe(fs.createWriteStream(process.cwd() + "/media/tags/" + cmd._id.toString()));
 }
 
+/** @type {Discord.Collection<string, import("../database/controllers/tag").tag>} */
+const tags = new Discord.Collection();
 
 /** @param {Discord.Message} msg */
 function runTag(msg) {
   const cmd = u.parse(msg);
   if (!msg.channel.isSendable() || !cmd) return;
 
-  const tag = tu.tags.get(cmd.command);
+  const tag = tags.get(cmd.command);
   if (!tag) return;
 
-  const encoded = tu.encodeTag(tag, msg);
+  const encoded = encodeTag(tag, msg);
   msg.reply({ ...encoded, allowedMentions: { parse: [] } }).then((/** @type {Discord.Message} */ m) => {
     if (typeof encoded === "string") u.clean(m);
   }).catch(u.noop);
+}
+
+
+/**
+ * @param {import("../database/controllers/tag").tag} tag
+ * @param {Discord.Message | null} msg
+ * @param {Discord.ChatInputCommandInteraction} [int]
+ */
+function encodeTag(tag, msg, int) {
+  let response = tag.response;
+  const user = msg?.inGuild() ? msg.member : int?.inCachedGuild() ? int.member : int?.user ?? msg?.author ?? null;
+  const origin = msg ?? int;
+  if (!user || !origin) return { content: "I couldn't process that command!" };
+  let target = msg?.mentions.members?.first() || msg?.mentions.users.first();
+  if (response) {
+    const randomChannels = origin.guild ? origin.guild.channels.cache.filter(c =>
+      c.isTextBased() && !c.isThread() && // normal text channel
+      !c.permissionOverwrites?.cache.get(origin.guild?.id ?? "")?.deny?.has("ViewChannel") // public channel
+    ).map(c => c.toString()) : ["Here"];
+
+    const regex = /<@random ?\[(.*?)\]>/gim;
+    if (regex.test(response)) {
+      response = response.replace(regex, (str) => u.rand(str.replace(regex, '$1').split('|')));
+    }
+
+    response = response
+      .replace(/<@channel>/ig, origin.channel?.toString() ?? "Here")
+      .replace(/<@randomchannel>/, u.rand(randomChannels) ?? origin.channel?.toString() ?? "Here")
+      .replace(/<@author>/ig, user.toString())
+      .replace(/<@authorname>/ig, user.displayName);
+
+    if ((/(<@target>)|(<@targetname>)/ig).test(response)) {
+      if (!origin.guild) target ??= origin.client.user;
+      if (!target) return { content: "You need to `@mention` a user with that command!" };
+      response = response
+        .replace(/<@target>/ig, target.toString())
+        .replace(/<@targetname>/ig, target.displayName);
+    }
+  }
+  return {
+    content: response ?? undefined,
+    files: tag.attachment ? [new u.Attachment(`./media/tags/${tag._id}`).setName(tag.attachment)] : [],
+    allowedMentions: { users: target ? [target.id, user.id] : [user.id] }
+  };
 }
 
 function contentModal(value = "") {
@@ -59,7 +104,7 @@ async function slashTagCreate(int) {
   // Get and validate input
   const name = int.options.getString('name', true).toLowerCase().replace(/[ \n]/g, "");
   const attachment = int.options.getAttachment('attachment');
-  if (tu.tags.has(name)) return int.reply({ content: `Looks like that tag already exists. Try </tag modify:${u.sf.commands.slashTag}> or </tag delete:${u.sf.commands.slashTag}> instead.`, ephemeral: true });
+  if (tags.has(name)) return int.reply({ content: `Looks like that tag already exists. Try </tag modify:${u.sf.commands.slashTag}> or </tag delete:${u.sf.commands.slashTag}> instead.`, ephemeral: true });
 
   await int.showModal(contentModal());
   const content = await int.awaitModalSubmit({ time: 5 * 60 * 1000, dispose: true }).catch(u.noop);
@@ -78,7 +123,7 @@ async function slashTagCreate(int) {
   if (!command) return content.editReply("I wasn't able to save that. Please try again later or with a different name.");
   if (attachment) saveAttachment(attachment, command);
 
-  tu.tags.set(command.tag, command);
+  tags.set(command.tag, command);
   const embed = u.embed({ author: int.member })
     .setTitle("Tag created")
     .setDescription(`${int.member} added the tag \`${name}\``);
@@ -98,7 +143,7 @@ async function slashTagCreate(int) {
 async function slashTagModify(int) {
   // get and validate inputs
   const name = int.options.getString('name', true).toLowerCase().replace(/[ \n]/g, "");
-  const currentTag = tu.tags.get(name);
+  const currentTag = tags.get(name);
   if (!currentTag) return int.reply({ content: `I couldn't find that tag.`, ephemeral: true });
 
   await int.showModal(contentModal(currentTag.response ?? ""));
@@ -120,7 +165,7 @@ async function slashTagModify(int) {
   if (!command) return content.editReply("I wasn't able to update that. Please try again later or contact a dev to see what went wrong.");
   if (attachment) saveAttachment(attachment, command);
 
-  tu.tags.set(command.tag, command);
+  tags.set(command.tag, command);
   const embed = u.embed({ author: int.member })
     .setTitle("Tag modified")
     .setDescription(`${int.member} modified the tag \`${name}\``);
@@ -148,11 +193,11 @@ async function slashTagModify(int) {
 async function slashTagDelete(int) {
   await int.deferReply({ ephemeral: true });
   const name = int.options.getString('name', true).toLowerCase().replace(/[ \n]/g, "");
-  if (!tu.tags.get(name)) return int.editReply(`Looks like that tag doesn't exist.`);
+  if (!tags.get(name)) return int.editReply(`Looks like that tag doesn't exist.`);
 
   const command = await u.db.tags.deleteTag(name);
   if (!command) return int.editReply("I wasn't able to delete that. Please try again later or contact a dev to see what went wrong.");
-  tu.tags.delete(name);
+  tags.delete(name);
 
   const embed = u.embed({ author: int.member })
     .setTitle("Tag Deleted")
@@ -195,7 +240,7 @@ async function slashTagValue(int) {
   await int.deferReply({ ephemeral: true });
 
   const name = int.options.getString('name', true).toLowerCase().replace(/[ \n]/g, "");
-  const tag = tu.tags.get(name);
+  const tag = tags.get(name);
   if (!tag) return int.editReply(`Looks like that tag doesn't exist.`);
 
   const embed = u.embed({ author: int.member })
@@ -223,7 +268,7 @@ const Module = new Augur.Module()
   },
   autocomplete: (int) => {
     const focusedValue = int.options.getFocused()?.toLowerCase();
-    const filtered = tu.tags.filter(tag => tag.tag.toLowerCase().includes(focusedValue));
+    const filtered = tags.filter(tag => tag.tag.toLowerCase().includes(focusedValue));
     return int.respond(filtered.map(choice => ({ name: choice.tag, value: choice.tag })).sort((a, b) => a.name.localeCompare(b.name)).slice(0, 24));
   }
 })
@@ -234,8 +279,9 @@ const Module = new Augur.Module()
 .setInit(async () => {
   try {
     const cmds = await u.db.tags.fetchAllTags();
-    cmds.forEach(c => tu.tags.set(c.tag.toLowerCase(), c));
+    cmds.forEach(c => tags.set(c.tag.toLowerCase(), c));
   } catch (error) { u.errorHandler(error, "Load Custom Tags"); }
-});
+})
+.addShared("tags.js", { tags, encodeTag });
 
 module.exports = Module;
